simple-kafka-dataflow-bigquery/requirements.txt
kafka-python
apache-beam
apache-beam[gcp]
Faker